08-28

+++ project_team3/worldFinder/src/main/java/org/worldfinder/domain/ScrapVO.java
@@ -1,5 +1,7 @@
 package org.worldfinder.domain;
 
+import java.sql.Date;
+
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -8,6 +10,10 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class ScrapVO {
-	private String u_writer;
-	private int up_idx;
+	private String u_writer;	// 현재 로그인한 사용자의 아이디 값
+	private long up_idx;
+	
+	//private String country;
+	//private String title;
+	//private Date reg_date;
 }
